+++ CHANGELOG.rst
@@ -4,5 +4,5 @@ Changelog
 1.0.0 - Unreleased
 ------------------
 
-- initial release
+- separated into plugin from devpi-server
   [fschulze (Florian Schulze)]